django_app/requirements.txt
Django==2.2

django-rq==2.4.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.6.0

Jinja2==2.10.1

joblib==0.15.1

jsonpickle
jsonschema


redis==3.5.3

requests==2.25.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1

rq==1.9.0
rq-scheduler==0.11.0

uWSGI==2.0.19.1

